Rewrite Euclidean Rhythm Generation (#2)
Replaces the existing Euclidean rhythm generation function with one that is accurate to the original Euclidean rhythms paper (Toussaint 2005).

It is an original (novel?) implementation of the algorithm described in the paper, which is more memory-efficient, CPU-efficient, and has a significantly smaller generated code size than the existing implementation. It does not use recursive code or non-integral numeric types. It functions by iteratively compressing the pattern into sub-sequences of growing complexity, and uses a string of bits to represent the working pattern of sub-sequences.