Separate input handling code from business logic (#3)
Using InputEvents struct to hold events that were received this cycle and can be processed. Rearranged a lot of core logic such that input processing happens first in the cycle. In addition to making the code easier to follow, this has the effect of reducing the number of global variables and decreasing UI latency slightly, since events are now handled on the cycle they are received in.